requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
bs4==0.0.1
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
conda==4.3.16
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.2
djangorestframework==3.12.4
fake-useragent==0.1.11
filelock==3.0.12
flower==0.9.7
humanize==3.4.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
lxml==4.6.3
MarkupSafe==1.1.1
packaging==20.9
Pillow==8.2.0
pipenv==2020.8.13
prometheus-client==0.8.0
prompt-toolkit==3.0.18
pycosat==0.6.3
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.4
vine==1.3.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
